Keezi walks can now be used. Tested
Tested Walking sequence renderer. It is possible to load it from pickle
or read the video sequence. Usually preprocessed pickles are very fast to load.
However since we load the entire database in memory it consumes about 75GB
of memory.

Now the normalization procedure is different. We feed unnormalized images
into tensorflow. We use tensorflow internal construct for speedy normalization.
This normalization is done in GPU.